requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-filter==22.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
idna==3.3
Markdown==3.4.1
numpy==1.23.1
openpyxl==3.0.10
pandas==1.4.3
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.10
